fix(cli): aspect help with no args should display root help (#1196)
This regressed in #993 when `aspect help flags-as-proto` support was
added.

Output is now correct again:

```
$ aspect help
Aspect CLI Pro is a better frontend for running bazel

Usage:
  aspect [command]

Common Bazel Commands:
  build              Build the specified targets
  cquery             Query the dependency graph, honoring configuration flags
  run                Build a single target and run it with the given arguments
  test               Build the specified targets and run all test targets among them
  version            Print the versions of Aspect CLI and Bazel

Commands only in Aspect CLI:
  configure          Update BUILD files for JavaScript, TypeScript, Go and Protobuf
  docs               Open documentation in the browser
  init               Create a new Bazel workspace
  outputs            Print paths to declared output files
  print              Print syntax elements from BUILD files
  support            Interactive, human-escalated support for Bazel problems

Custom Commands from Plugins:

Other Bazel Built-in Commands:
  analyze-profile    Analyze build profile data
  aquery             Query the action graph
  canonicalize-flags Present a list of bazel options in a canonical form
  clean              Remove the output tree
  coverage           Same as 'test', but also generates a code coverage report.
  fetch              Fetch external repositories that are prerequisites to the targets
  info               Display runtime info about the bazel server
  license            Prints the license of this software.
  query              Query the dependency graph, ignoring configuration flags
  shutdown           Stop the bazel server

Additional Commands:
  completion         Generate the autocompletion script for the specified shell
  help

Flags:
      --aspect:config string   User-specified Aspect CLI config file. /dev/null indicates that all further --aspect:config flags will be ignored.
      --aspect:interactive     Interactive mode (e.g. prompts for user input) (default true)

Additional help topics:
  aspect info-keys          Displays a list of keys used by the info command.
  aspect tags               Conventions for tags which are special.
  aspect target-syntax      Explains the syntax for specifying targets.

Use "aspect [command] --help" for more information about a command.
```

GitOrigin-RevId: 5546d8eb5c5d296aa6bc3152611bedca9e04f35f
