+++ cmd/aspect/root/BUILD.bazel
@@ -35,6 +35,7 @@ go_library(
         "//cmd/aspect/sync",
         "//cmd/aspect/test",
         "//cmd/aspect/version",
+        "//docs/help/topics",
         "//pkg/aspect/root/flags",
         "//pkg/ioutils",
         "//pkg/plugin/system",
